requirements.txt
speechrecognition
pyttsx3
pyaudio
pywhatkit
wikipedia
pyjokesc